% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nmf_modules.R
\name{add_nmf_modules}
\alias{add_nmf_modules}
\title{add_nmf_modules}
\usage{
add_nmf_modules(seurat_obj, nmf_results)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{nmf_results}{nmf results, ouput from perform_nmf.}
}
\value{
Seurat object.
}
\description{
Add NMF module scores to Seurat object metadata.
}
