% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_FindMarkers_human.R,
%   R/convert_FindMarkers_human_old.R
\name{convert_FindMarkers_human}
\alias{convert_FindMarkers_human}
\title{convert_FindMarkers_human}
\usage{
convert_FindMarkers_human(
  fish_genelist,
  create_ranked_vector = T,
  remove_ribosomal_genes = T
)

convert_FindMarkers_human(
  fish_genelist,
  create_ranked_vector = T,
  remove_ribosomal_genes = T
)
}
\arguments{
\item{fish_genelist}{FindMarkers results dataframe.}

\item{create_ranked_vector}{if T, will convert the output into a ranked vector with human genes as name and avg_logFC as values, to use for GSEA.}

\item{remove_ribosomal_genes}{if T, will remove anything starting with rps or rpl from the gene list}

\item{genelist}{FindMarkers output dataframe. Needs to have a column with the gene names called "gene" to proceed with this function.}
}
\value{
ranked vector or dataframe with converted human genes.

if create_ranked_vector = T, a ranked vector with human genes as name and avg_logFC as values, to use for GSEA. if F, a dataframe.
}
\description{
convert the results of FindMarkers (dataframe) to human genes and create ranked vector for GSEA. * Improved version!

convert the output of the Seurat function FindMarkers from fish genes to human.
}
