% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_seurat_obj_to_human.R
\name{convert_seurat_obj_to_human}
\alias{convert_seurat_obj_to_human}
\title{convert_seurat_obj_to_human}
\usage{
convert_seurat_obj_to_human(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat object.}
}
\value{
Seurat object with human genes instead of fish. Image data will be preserved if applying this function to Visium data.
}
\description{
convert either Visium or scRNA-seq Seurat object from fish genes to human.
}
