% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment_nmf.R
\name{go_enrichment_nmf}
\alias{go_enrichment_nmf}
\title{go_enrichment_nmf}
\usage{
go_enrichment_nmf(
  nmf_results,
  num_genes = 150,
  n_terms_plot = 10,
  species = "fish"
)
}
\arguments{
\item{nmf_results}{nmf_results, output from perform_nmf.}

\item{num_genes}{number of top scoring genes per factor to be used for GO analysis.}

\item{n_terms_plot}{number of GO terms to be plotted per factor.}

\item{species}{species to use for GO enrichment. must be one of 'fish', 'zebrafish', or 'human'.}
}
\value{
results as list including dataframes with all results, and plots.
}
\description{
Perform GO term enrichment on NMF results using GProfiler. Both GO BP and GO CC enrichment will be run.
}
