% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_waterfall_plot.R
\name{gsea_waterfall_plot}
\alias{gsea_waterfall_plot}
\title{gsea_waterfall_plot}
\usage{
gsea_waterfall_plot(
  gsea_results,
  pathways_to_highlight,
  labels_for_plot = NULL,
  n_terms = 250,
  transparent = T,
  col_label = NULL
)
}
\arguments{
\item{gsea_results}{your GSEA results dataframe, should have one column called "pathway" and another called "NES", additional columns are fine too.}

\item{pathways_to_highlight}{a character vector with the exact names of the pathways you want highlighted on your waterfall plot in red circles.}

\item{labels_for_plot}{optional: a character vector with the exact names of the pathways you want labelled as text on your plot.}

\item{n_terms}{the number of terms to be plotted.}

\item{transparent}{if you want the non-labelled points to be transparent - looks better but causes issues with exporting.}

\item{col_label}{colour to label points of interest with.}
}
\value{
double waterfall plot.
}
\description{
double waterfall plot of the chosen top and bottom pathways by NES from GSEA. Optional labelling of pathways of interest.
}
