% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invasion_assay_plotting_function.R
\name{invasion_assay_plotting_function}
\alias{invasion_assay_plotting_function}
\title{invasion_assay_plotting_function}
\usage{
invasion_assay_plotting_function(plot_data, sep_by_replicates = F)
}
\arguments{
\item{plot_data}{data frame with normalized results.}

\item{sep_by_replicates}{Plot each replicate separately?}
}
\value{
Plots with results from the experiment. If sep_by_replicates = F and there are multiple replicates, they will be labelled with different shapes on the plot.
}
\description{
Plot normalized data from invasion assay.
}
