% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_bayesspace_feature_plot.R
\name{nice_bayesspace_feature_plot}
\alias{nice_bayesspace_feature_plot}
\title{nice_bayesspace_feature_plot}
\usage{
nice_bayesspace_feature_plot(
  enhanced_obj,
  normal_obj = NULL,
  features,
  cols = brewer.rdpu(n = 100),
  show_comparison = F,
  show_title = F,
  plot_enhanced = T,
  n_col = NULL
)
}
\arguments{
\item{enhanced_obj}{Name of enhanced BayesSpace object.}

\item{normal_obj}{Optional: name of Seurat object to show comparison.}

\item{features}{Genes to plot expression of.}

\item{cols}{Colormap for plotting.}

\item{show_comparison}{if T, will plot both "normal" expression in Seurat object as well as enhanced expression.}

\item{show_title}{show titles on plots?}

\item{plot_enhanced}{if F, will only plot the Seurat expression.}

\item{n_col}{number of columns for final plot.}
}
\value{
plots.
}
\description{
Plot enhanced expression of a given gene resulting from BayesSpace analysis.
}
