% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_dim_plot.R, R/nice_dim_plot_old.R
\name{nice_dim_plot}
\alias{nice_dim_plot}
\title{nice_dim_plot}
\usage{
nice_dim_plot(
  seurat_obj,
  group_by = NULL,
  cols = NULL,
  pt_size = 1.3,
  label = T,
  reduction = "umap",
  dims_plot = 1:2,
  n_col = NULL
)

nice_dim_plot(
  seurat_obj,
  group_by = NULL,
  cols = NULL,
  pt_size = 1.3,
  label = T,
  reduction = "umap",
  dims_plot = 1:2,
  n_col = NULL
)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{group_by}{optional: what to colour the points by, usually a column in the Seurat object metadata. Otherwise will color by the default ident.}

\item{cols}{optional: colours to label the grouped points by.}

\item{pt_size}{point size.}

\item{label}{label groups with text on plot?}

\item{reduction}{either "umap" or "pca".}

\item{dims_plot}{dimensions to plot.}

\item{n_col}{number of columns, only relevant if length of group_by > 1}
}
\value{
UMAP or PCA plot.

UMAP or PCA plot.
}
\description{
nicer looking version of the Seurat function DimPlot.

nicer looking version of the Seurat function DimPlot.
}
