% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet_wrapper_MVH.R
\name{nichenet_wrapper_MVH}
\alias{nichenet_wrapper_MVH}
\title{nichenet_wrapper_MVH}
\usage{
nichenet_wrapper_MVH(
  seurat_obj,
  sender_ident,
  receiver_ident,
  ligand_target_cutoff = 0.5
)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{sender_ident}{Name of cluster you want to be considered as the "sender".}

\item{receiver_ident}{Name of cluster you want to be considered as the "receiver".}

\item{ligand_target_cutoff}{Cutoff for ligand-target analysis.}
}
\value{
Nichenet results plots.
}
\description{
perform Nichenet analysis on clusters of interest.
}
