% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_nmf.R
\name{perform_nmf}
\alias{perform_nmf}
\title{perform_nmf}
\usage{
perform_nmf(
  expression_matrix,
  rank = 20,
  save_NMF_results = F,
  filename = NULL
)
}
\arguments{
\item{expression_matrix}{For scRNA-seq or Visium, typically the output from GetAssayData for your cluster/cell type of interest.}

\item{rank}{Rank used for NMF, can be determined using NMFestimaterank.}

\item{save_NMF_results}{if T, will save NMF function output as .R file. This can be useful since the function takes several hours to run.}

\item{filename}{if save_NMF_results = T, filename that the NMF results will be saved as.}
}
\value{
a list containing the NMF results matrix, coefficients, scores, and modules.
}
\description{
perform NMF on a matrix of expression data.
}
