% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ccle_data_MVH.R
\name{plot_ccle_data_MVH}
\alias{plot_ccle_data_MVH}
\title{plot_ccle_data_MVH}
\usage{
plot_ccle_data_MVH(ccle_data, genes, cell_line_id = "A375")
}
\arguments{
\item{ccle_data}{CCLE log2 RSEM data, with genes as rownames as HGNC ID.}

\item{genes}{genelist to plot.}

\item{cell_line_id}{Cell line you want to plot. If A375 or SKMEL28, can just put those names. If anything else, need to put DepMap cell line ID.}
}
\value{
bargraph of normalized expression of the given genes.
}
\description{
Plot RNA-seq data from CCLE/DepMap repository. ccle_data.R, aka CCLE_expression.csv, and ccle_sample_info.R, aka sample_info.csv, must be loaded first for this function to work. Both files can be downloaded from: https://depmap.org/portal/download/
}
