% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression_bulkRNAseq.R
\name{plot_expression_bulkRNAseq}
\alias{plot_expression_bulkRNAseq}
\title{plot_expression_bulkRNAseq}
\usage{
plot_expression_bulkRNAseq(genes, plot_counts = F, plot_norm_expression = T)
}
\arguments{
\item{genes}{gene to plot expression of.}

\item{plot_counts}{plot expression of raw counts?}

\item{plot_norm_expression}{plot normalized expression?}
}
\value{
plot.
}
\description{
Plot expression - counts or normalized - of a given gene from bulk RNA-seq expression data.
}
