% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_per_cluster.R
\name{plot_per_cluster}
\alias{plot_per_cluster}
\title{plot_per_cluster}
\usage{
plot_per_cluster(seurat_obj, group_clusters, metadata_col_to_plot, cols = NULL)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{group_clusters}{the metadata column you want to group the clusters by, e.g cell type.}

\item{metadata_col_to_plot}{the metadata column you want to plot the proportion of cells per cluster for, e.g cell cycle phase.}

\item{cols}{optional: colours to plot the diff levels of metadata_col_to_plot by.}
}
\value{
violin plot.
}
\description{
plot the relative proportion of cells per cluster grouped by a categorical variable. for example, percentage of cells per cluster in each of the 3 cell cycle phases.
}
