% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cyquant_plate.R
\name{process_cyquant_plate}
\alias{process_cyquant_plate}
\title{process_cyquant_plate}
\usage{
process_cyquant_plate(path_to_cyquant_plate)
}
\arguments{
\item{path_to_cyquant_plate}{Path to the excel file with the data from the plate reader.}
}
\value{
Plots with results from the experiment.
}
\description{
Process and plot data from a 96 well plate used with Cyquant assay. This is only set up to use with a specific assay and plate layout - not for general use.
}
