% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea_gobp.R
\name{run_gsea_gobp}
\alias{run_gsea_gobp}
\title{run_gsea_gobp}
\usage{
run_gsea_gobp(marker_list, filter_pval = T)
}
\arguments{
\item{marker_list}{FindMarkers output, not converted to human.}

\item{filter_pval}{if T, filters FindMarkers results to remove p > 0.05.}
}
\value{
fGSEA results as dataframe.
}
\description{
Run fGSEA on GO biological processes pathway set.
}
