% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_nmf.R
\name{perform_nmf}
\alias{perform_nmf}
\title{perform_nmf}
\usage{
perform_nmf(
  expression_matrix,
  rank = 20,
  save_NMF_results = F,
  filename = NULL,
  use_fast_method = F
)
}
\arguments{
\item{expression_matrix}{Expression matrix, commonly variable genes, for cell type of interest.}

\item{rank}{Rank for NMF, default = 20. Can be a vector of possible ranks but can only use multiple ranks if use_fast_method = F.}

\item{save_NMF_results}{Save NMF results matrix in working dir?}

\item{filename}{Filename to save NMF results as.}

\item{use_fast_method}{If T, will use the fast NMF method implemented in RcppML. Can only be used for a single rank.}
}
\value{
NMF results list with matrix, modules, scores and coef.
}
\description{
Perform NMF and create gene modules from expression matrix. Slightly modified code from Reuben and Dalia.
}
